Fixes #37834 - Set Firmware selection correct
Without this PR, its is currently really broken. Without this PR, it
does always! show 'Automatic' - except while editing hosts. For
new/edit of Compute Profile or in the Create Host page - always
'Automatic' is shown. A user would never know what is really configured.
This is because of [1].

If you create a host later, 'Automatic' is always displayed, regardless
of which firmware is selected in the Compute Profile.

This change makes sure, that the selected firmware on the Compute
Profile is the one which was saved previously and represents the
value from the database while editing a existing Compute Profile.

The `process_firmware_attributes` is modified so that the methods
`resolve_automatic_firmware` and `normalize_firmware_type` are only used
if a host is created on the compute resource.

[1] https://github.com/theforeman/foreman/pull/10321/files#diff-c88b347f9af46e109987241498e9db3c776cc76433ffe3ba6f58ae1d8a74b9adR11