Fixes #38418 - Pin spring to 4.2.1 in development
Spring v4.3.0, in combination with Ruby 3.0.7 and Rails 7,
is causing an error that blocks the Rails start.

Error: method_missing: undefined method root
for Rails::Application:Class (NoMethodError)

One solution could be to upgrade Ruby to 3.1.7,
but that doesn't work on the latest Fedora 42 [1].

And upgrading to 3.2.z is not feasible due to some gems
limiting Ruby versions to < 3.2 (safe-render, I think).

External issues:
[0] https://github.com/rails/spring/issues/737
[1] https://github.com/rbenv/ruby-build/discussions/2529