Fixes #38732 - Deactivate /etc/grub.d/30_uefi-firmware in Debian 13
This GRUB2 configuration snippet contains the line `fwsetup
--is-supported`, where the parameter `--is-supported` is not understood
by older versions of GRUB2.  This may include GRUB2 from Foreman
proxies.

This GRUB2 configuration snippet adds a menu entry in GRUB2 to enter the
BIOS screen that is negligible.

To prevent the file from possibly being recreated by a package update
(grub-common), we overwrite the content instead of removing it.